@@ -1,136 +1,136 @@
 {
-  "name": "zeus",
-  "version": "0.0.1",
-  "description": "Self-order Point of Sale solution, powered by the Lightning Network",
-  "license": "MIT",
-  "cacheDirectories": [
-    "node_modules"
-  ],
-  "dependencies": {
-    "@angular/common": "6.1.0",
-    "@angular/compiler": "6.1.0",
-    "@angular/core": "6.1.0",
-    "@angular/forms": "6.1.0",
-    "@angular/platform-browser": "6.1.0",
-    "@angular/platform-browser-dynamic": "6.1.0",
-    "@angular/router": "6.1.0",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.8",
-    "@fortawesome/free-brands-svg-icons": "^5.5.0",
-    "@fortawesome/free-solid-svg-icons": "^5.5.0",
-    "@ng-bootstrap/ng-bootstrap": "2.2.1",
-    "angularx-qrcode": "^1.2.2",
-    "bootstrap": "^4.3.1",
-    "core-js": "2.5.7",
-    "jquery": "^3.4.1",
-    "light-it-up": "^1.0.0",
-    "lodash": "^4.17.10",
-    "moment": "2.22.2",
-    "ng-jhipster": "0.5.4",
-    "ngx-cookie": "2.0.1",
-    "ngx-infinite-scroll": "0.5.1",
-    "ngx-webstorage": "2.0.1",
-    "reflect-metadata": "0.1.12",
-    "rxjs": "6.1.0",
-    "rxjs-compat": "6.1.0",
-    "sockjs-client": "1.1.4",
-    "swagger-ui": "2.2.10",
-    "tableexport": "^5.0.2",
-    "tslib": "1.9.3",
-    "webstomp-client": "1.2.0",
-    "zone.js": "0.8.26"
-  },
-  "devDependencies": {
-    "@angular/cli": "6.1.2",
-    "@angular/compiler-cli": "6.1.0",
-    "@ngtools/webpack": "6.0.0",
-    "@types/jest": "22.2.3",
-    "@types/node": "9.4.7",
-    "angular-router-loader": "0.8.5",
-    "angular2-template-loader": "0.6.2",
-    "browser-sync": "2.24.4",
-    "browser-sync-webpack-plugin": "2.2.2",
-    "cache-loader": "1.2.2",
-    "codelyzer": "4.2.1",
-    "copy-webpack-plugin": "4.5.1",
-    "css-loader": "0.28.10",
-    "exports-loader": "0.7.0",
-    "extract-text-webpack-plugin": "4.0.0-beta.0",
-    "file-loader": "1.1.11",
-    "fork-ts-checker-webpack-plugin": "0.4.1",
-    "friendly-errors-webpack-plugin": "1.7.0",
-    "generator-jhipster": "5.2.0",
-    "html-loader": "0.5.5",
-    "html-webpack-plugin": "3.2.0",
-    "husky": "0.14.3",
-    "jest": "22.4.3",
-    "jest-junit": "5.1.0",
-    "jest-preset-angular": "5.2.2",
-    "jest-sonar-reporter": "2.0.0",
-    "lint-staged": "7.0.0",
-    "moment-locales-webpack-plugin": "1.0.5",
-    "node-sass": "4.9.0",
-    "postcss-loader": "2.1.1",
-    "prettier": "1.11.1",
-    "proxy-middleware": "0.15.0",
-    "raw-loader": "0.5.1",
-    "rimraf": "2.6.1",
-    "sass-loader": "7.0.1",
-    "simple-progress-webpack-plugin": "1.1.2",
-    "style-loader": "0.20.3",
-    "tapable": "1.0.0",
-    "terser-webpack-plugin": "1.0.0",
-    "thread-loader": "1.1.5",
-    "to-string-loader": "1.1.5",
-    "ts-loader": "4.0.1",
-    "tslint": "5.9.1",
-    "tslint-config-prettier": "1.9.0",
-    "tslint-loader": "3.6.0",
-    "typescript": "2.7.2",
-    "webpack": "4.8.0",
-    "webpack-cli": "2.1.3",
-    "webpack-dev-server": "^3.1.11",
-    "webpack-merge": "4.1.2",
-    "webpack-notifier": "1.6.0",
-    "webpack-visualizer-plugin": "0.1.11",
-    "workbox-webpack-plugin": "3.2.0",
-    "write-file-webpack-plugin": "4.2.0"
-  },
-  "engines": {
-    "yarn": ">=1.3.2"
-  },
-  "lint-staged": {
-    "src/**/*.{ts,css,scss}": [
-      "prettier --write",
-      "git add"
-    ]
-  },
-  "scripts": {
-    "precommit": "lint-staged",
-    "prettier:format": "yarn prettier --write 'src/**/*.{ts,css,scss}'",
-    "lint": "tslint --project tsconfig.json -e 'node_modules/**'",
-    "lint:fix": "yarn run lint --fix",
-    "ngc": "ngc -p tsconfig-aot.json",
-    "cleanup": "rimraf build/{aot,www}",
-    "clean-www": "rimraf build//www/app/{src,build/}",
-    "start": "yarn run webpack:dev",
-    "serve": "yarn run start",
-    "build": "yarn run webpack:prod",
-    "test": "yarn run lint && jest --coverage --logHeapUsage -w=2 --config test/jest.conf.js",
-    "test:watch": "yarn test --watch --clearCache",
-    "webpack:dev": "yarn run webpack-dev-server --config webpack/webpack.dev.js --inline --hot --host=0.0.0.0 --port=9060 --watch-content-base --env.stats=minimal",
-    "webpack:dev-cold": "yarn run webpack-dev-server --config webpack/webpack.dev.js --inline --host=0.0.0.0 --port=9060 --watch-content-base --env.stats=minimal",
-    "webpack:dev-verbose": "yarn run webpack-dev-server --config webpack/webpack.dev.js --inline --hot --port=9060 --watch-content-base --profile --progress --env.stats=normal",
-    "webpack:build:main": "yarn run webpack --config webpack/webpack.dev.js --env.stats=normal",
-    "webpack:build": "yarn run cleanup && yarn run webpack:build:main",
-    "webpack:prod:main": "yarn run webpack --config webpack/webpack.prod.js --profile",
-    "webpack:prod": "yarn run cleanup && yarn run webpack:prod:main && yarn run clean-www",
-    "webpack:test": "yarn run test",
-    "webpack-dev-server": "node --max_old_space_size=4096 node_modules/webpack-dev-server/bin/webpack-dev-server.js",
-    "webpack": "node --max_old_space_size=4096 node_modules/webpack/bin/webpack.js"
-  },
-  "jestSonar": {
-    "reportPath": "build/test-results/jest",
-    "reportFile": "TESTS-results-sonar.xml"
-  }
+    "name": "zeus",
+    "version": "1.0.0",
+    "description": "Self-order Point of Sale solution, powered by the Lightning Network",
+    "license": "MIT",
+    "cacheDirectories": [
+        "node_modules"
+    ],
+    "dependencies": {
+        "@angular/common": "7.2.12",
+        "@angular/compiler": "7.2.12",
+        "@angular/core": "7.2.12",
+        "@angular/forms": "7.2.12",
+        "@angular/platform-browser": "7.2.12",
+        "@angular/platform-browser-dynamic": "7.2.12",
+        "@angular/router": "7.2.12",
+        "@fortawesome/angular-fontawesome": "0.3.0",
+        "@fortawesome/fontawesome-svg-core": "1.2.17",
+        "@fortawesome/free-brands-svg-icons": "^5.8.2",
+        "@fortawesome/free-solid-svg-icons": "5.8.1",
+        "@ng-bootstrap/ng-bootstrap": "4.1.1",
+        "@ngx-translate/core": "11.0.1",
+        "@ngx-translate/http-loader": "4.0.0",
+        "angularx-qrcode": "^1.2.2",
+        "bootstrap": "4.3.1",
+        "core-js": "2.6.5",
+        "jquery": "^3.4.1",
+        "light-it-up": "^1.0.0",
+        "lodash": "^4.17.10",
+        "moment": "2.24.0",
+        "ng-jhipster": "0.9.3",
+        "ngx-cookie": "4.0.2",
+        "ngx-infinite-scroll": "7.1.0",
+        "ngx-webstorage": "3.0.2",
+        "rxjs": "6.4.0",
+        "swagger-ui": "2.2.10",
+        "sockjs-client": "1.3.0",
+        "tableexport": "^5.0.2",
+        "webstomp-client": "1.2.6",
+        "tslib": "1.9.3",
+        "zone.js": "0.8.29"
+    },
+    "devDependencies": {
+        "@angular/cli": "7.3.8",
+        "@angular/compiler-cli": "7.2.12",
+        "@ngtools/webpack": "7.3.8",
+        "@types/jest": "24.0.11",
+        "@types/node": "10.12.27",
+        "angular-router-loader": "0.8.5",
+        "angular2-template-loader": "0.6.2",
+        "autoprefixer": "9.5.0",
+        "base-href-webpack-plugin": "2.0.0",
+        "browser-sync": "2.26.5",
+        "browser-sync-webpack-plugin": "2.2.2",
+        "autoprefixer": "9.4.7",
+        "cache-loader": "2.0.1",
+        "codelyzer": "5.0.0",
+        "copy-webpack-plugin": "5.0.2",
+        "css-loader": "2.1.1",
+        "file-loader": "3.0.1",
+        "fork-ts-checker-webpack-plugin": "1.0.1",
+        "friendly-errors-webpack-plugin": "1.7.0",
+        "generator-jhipster": "6.0.1",
+        "html-loader": "0.5.5",
+        "html-webpack-plugin": "3.2.0",
+        "husky": "1.3.1",
+        "jest": "24.7.1",
+        "jest-junit": "6.3.0",
+        "jest-preset-angular": "7.0.1",
+        "jest-sonar-reporter": "2.0.0",
+        "lint-staged": "8.1.5",
+        "mini-css-extract-plugin": "0.5.0",
+        "moment-locales-webpack-plugin": "1.0.7",
+        "optimize-css-assets-webpack-plugin": "5.0.1",
+        "postcss-loader": "3.0.0",
+        "prettier": "1.16.4",
+        "reflect-metadata": "0.1.13",
+        "rimraf": "2.6.3",
+        "sass": "1.17.4",
+        "sass-loader": "7.1.0",
+        "simple-progress-webpack-plugin": "1.1.2",
+        "style-loader": "0.23.1",
+        "terser-webpack-plugin": "1.2.3",
+        "thread-loader": "2.1.2",
+        "to-string-loader": "1.1.5",
+        "ts-loader": "5.3.3",
+        "tslint": "5.15.0",
+        "tslint-config-prettier": "1.18.0",
+        "tslint-loader": "3.5.4",
+        "typescript": "3.2.4",
+        "webpack": "4.29.6",
+        "webpack-cli": "3.3.0",
+        "webpack-dev-server": "3.2.1",
+        "webpack-merge": "4.2.1",
+        "webpack-notifier": "1.7.0",
+        "webpack-visualizer-plugin": "0.1.11",
+        "workbox-webpack-plugin": "3.6.3",
+        "write-file-webpack-plugin": "4.5.0"
+    },
+    "engines": {
+        "node": ">=8.9.0"
+    },
+    "lint-staged": {
+        "{,src/**/}*.{md,json,ts,css,scss,yml}": [
+            "prettier --write",
+            "git add"
+        ]
+    },
+    "scripts": {
+        "prettier:format": "prettier --write \"{,src/**/}*.{md,json,ts,css,scss,yml}\"",
+        "lint": "tslint --project tsconfig.json -e 'node_modules/**'",
+        "lint:fix": "npm run lint -- --fix",
+        "ngc": "ngc -p tsconfig-aot.json",
+        "cleanup": "rimraf build/resources/main/static/ build/resources/main/aot",
+        "clean-www": "rimraf build/resources/main/static/app/{src,build/}",
+        "start": "npm run webpack:dev",
+        "start-tls": "npm run webpack:dev -- --env.tls",
+        "serve": "npm run start",
+        "build": "npm run webpack:prod",
+        "test": "npm run lint && jest --coverage --logHeapUsage -w=2 --config src/test/javascript/jest.conf.js",
+        "test:watch": "npm run test -- --watch",
+        "webpack:dev": "npm run webpack-dev-server -- --config webpack/webpack.dev.js --inline --hot --port=9060 --watch-content-base --env.stats=minimal",
+        "webpack:dev-verbose": "npm run webpack-dev-server -- --config webpack/webpack.dev.js --inline --hot --port=9060 --watch-content-base --profile --progress --env.stats=normal",
+        "webpack:build:main": "npm run webpack -- --config webpack/webpack.dev.js --env.stats=minimal",
+        "webpack:build": "npm run cleanup && npm run webpack:build:main",
+        "webpack:prod:main": "npm run webpack -- --config webpack/webpack.prod.js --profile",
+        "webpack:prod": "npm run cleanup && npm run webpack:prod:main && npm run clean-www",
+        "webpack:test": "npm run test",
+        "webpack-dev-server": "node --max_old_space_size=4096 node_modules/webpack-dev-server/bin/webpack-dev-server.js",
+        "webpack": "node --max_old_space_size=4096 node_modules/webpack/bin/webpack.js"
+    },
+    "jestSonar": {
+        "reportPath": "build/test-results/jest",
+        "reportFile": "TESTS-results-sonar.xml"
+    }
 }
