@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { JhiEventManager } from 'ng-jhipster';
-import { Account, LoginModalService, Principal } from 'app/core';
 import { HomeService } from 'app/admin/home/home.service';
+import { Account, AccountService, LoginModalService } from 'app/core';
 
 export const KEY_SHOP_ACTIVE = 'shop.active';
 export const KEY_SHOP_OVERRIDE = 'shop.override';
@@ -18,14 +18,14 @@ export class JhiHomeComponent implements OnInit {
     shopOverrideConfig: any;
 
     constructor(
-        private principal: Principal,
+        private accountService: AccountService,
         private loginModalService: LoginModalService,
         private eventManager: JhiEventManager,
         private homeService: HomeService
     ) {}
 
     ngOnInit() {
-        this.principal.identity().then(account => {
+        this.accountService.identity().then((account: Account) => {
             this.account = account;
         });
         this.registerAuthenticationSuccess();
@@ -51,7 +51,7 @@ export class JhiHomeComponent implements OnInit {
 
     registerAuthenticationSuccess() {
         this.eventManager.subscribe('authenticationSuccess', message => {
-            this.principal.identity().then(account => {
+            this.accountService.identity().then(account => {
                 this.account = account;
             });
             this.loadConfig();
@@ -59,7 +59,7 @@ export class JhiHomeComponent implements OnInit {
     }
 
     isAuthenticated() {
-        return this.principal.isAuthenticated();
+        return this.accountService.isAuthenticated();
     }
 
     login() {
