@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
-import { combineLatest } from 'rxjs/index';
+import { combineLatest, Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Rx';
 import { getServerUrl } from 'app/app.constants';
 import { map } from 'rxjs/internal/operators';
 
