@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
-import { JhiMetricsMonitoringModalComponent } from './metrics-modal.component';
 import { JhiMetricsService } from './metrics.service';
 
 @Component({
@@ -10,8 +9,7 @@ import { JhiMetricsService } from './metrics.service';
 })
 export class JhiMetricsMonitoringComponent implements OnInit {
     metrics: any = {};
-    cachesStats: any = {};
-    servicesStats: any = {};
+    threadData: any = {};
     updatingMetrics = true;
     JCACHE_KEY: string;
 
@@ -27,51 +25,18 @@ export class JhiMetricsMonitoringComponent implements OnInit {
         this.updatingMetrics = true;
         this.metricsService.getMetrics().subscribe(metrics => {
             this.metrics = metrics;
-            this.updatingMetrics = false;
-            this.servicesStats = {};
-            this.cachesStats = {};
-            Object.keys(metrics.timers).forEach(key => {
-                const value = metrics.timers[key];
-                if (key.includes('web.rest') || key.includes('service')) {
-                    this.servicesStats[key] = value;
-                }
-            });
-            Object.keys(metrics.gauges).forEach(key => {
-                if (key.includes('jcache.statistics')) {
-                    const value = metrics.gauges[key].value;
-                    // remove gets or puts
-                    const index = key.lastIndexOf('.');
-                    const newKey = key.substr(0, index);
-
-                    // Keep the name of the domain
-                    this.cachesStats[newKey] = {
-                        name: this.JCACHE_KEY.length,
-                        value
-                    };
-                }
+            this.metricsService.threadDump().subscribe(data => {
+                this.threadData = data.threads;
+                this.updatingMetrics = false;
             });
         });
     }
 
-    refreshThreadDumpData() {
-        this.metricsService.threadDump().subscribe(data => {
-            const modalRef = this.modalService.open(JhiMetricsMonitoringModalComponent, { size: 'lg' });
-            modalRef.componentInstance.threadDump = data.threads;
-            modalRef.result.then(
-                result => {
-                    // Left blank intentionally, nothing to do here
-                },
-                reason => {
-                    // Left blank intentionally, nothing to do here
-                }
-            );
-        });
+    isObjectExisting(metrics: any, key: string) {
+        return metrics && metrics[key];
     }
 
-    filterNaN(input) {
-        if (isNaN(input)) {
-            return 0;
-        }
-        return input;
+    isObjectExistingAndNotEmpty(metrics: any, key: string) {
+        return this.isObjectExisting(metrics, key) && JSON.stringify(metrics[key]) !== '{}';
     }
 }
