@@ -6,7 +6,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { JhiAlertService, JhiEventManager, JhiParseLinks } from 'ng-jhipster';
 
 import { ITEMS_PER_PAGE } from 'app/shared';
-import { Principal, User, UserService } from 'app/core';
+import { Account, AccountService, User, UserService } from 'app/core';
 
 @Component({
     selector: 'jhi-user-mgmt',
@@ -30,7 +30,7 @@ export class UserMgmtComponent implements OnInit, OnDestroy {
     constructor(
         private userService: UserService,
         private alertService: JhiAlertService,
-        private principal: Principal,
+        private accountService: AccountService,
         private parseLinks: JhiParseLinks,
         private activatedRoute: ActivatedRoute,
         private router: Router,
@@ -47,7 +47,7 @@ export class UserMgmtComponent implements OnInit, OnDestroy {
     }
 
     ngOnInit() {
-        this.principal.identity().then(account => {
+        this.accountService.identity().then((account: Account) => {
             this.currentAccount = account;
             this.loadAll();
             this.registerChangeInUsers();
