@@ -1,11 +1,12 @@
-import { Injector } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 import { LoginService } from 'app/core/login/login.service';
 
+@Injectable()
 export class AuthExpiredInterceptor implements HttpInterceptor {
-    constructor(private injector: Injector) {}
+    constructor(private loginService: LoginService) {}
 
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         return next.handle(request).pipe(
@@ -14,8 +15,7 @@ export class AuthExpiredInterceptor implements HttpInterceptor {
                 (err: any) => {
                     if (err instanceof HttpErrorResponse) {
                         if (err.status === 401) {
-                            const loginService: LoginService = this.injector.get(LoginService);
-                            loginService.logout();
+                            this.loginService.logout();
                         }
                     }
                 }
