@@ -1,8 +1,10 @@
+import { Injectable } from '@angular/core';
 import { JhiEventManager } from 'ng-jhipster';
 import { HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
+@Injectable()
 export class ErrorHandlerInterceptor implements HttpInterceptor {
     constructor(private eventManager: JhiEventManager) {}
 
@@ -13,9 +15,7 @@ export class ErrorHandlerInterceptor implements HttpInterceptor {
                 (err: any) => {
                     if (err instanceof HttpErrorResponse) {
                         if (!(err.status === 401 && (err.message === '' || (err.url && err.url.includes('/api/account'))))) {
-                            if (this.eventManager !== undefined) {
-                                this.eventManager.broadcast({ name: 'zeusApp.httpError', content: err });
-                            }
+                            this.eventManager.broadcast({ name: 'zeusApp.httpError', content: err });
                         }
                     }
                 }
