@@ -2,7 +2,10 @@ import { LOCALE_ID, NgModule } from '@angular/core';
 import { DatePipe, registerLocaleData } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { Title } from '@angular/platform-browser';
-import locale from '@angular/common/locales/de-CH';
+import locale from '@angular/common/locales/de';
+
+import { NgbDateAdapter } from '@ng-bootstrap/ng-bootstrap';
+import { NgbDateMomentAdapter } from 'app/shared/util/datepicker-adapter';
 
 @NgModule({
     imports: [HttpClientModule],
@@ -14,6 +17,7 @@ import locale from '@angular/common/locales/de-CH';
             provide: LOCALE_ID,
             useValue: 'de-CH'
         },
+        { provide: NgbDateAdapter, useClass: NgbDateMomentAdapter },
         DatePipe
     ]
 })
