@@ -1,12 +1,16 @@
 import { Injectable } from '@angular/core';
 
-import { Principal } from '../auth/principal.service';
-import { AuthServerProvider } from '../auth/auth-jwt.service';
-import { JhiTrackerService } from '../tracker/tracker.service';
+import { AccountService } from 'app/core/auth/account.service';
+import { AuthServerProvider } from 'app/core/auth/auth-jwt.service';
+import { JhiTrackerService } from 'app/core/tracker/tracker.service';
 
 @Injectable({ providedIn: 'root' })
 export class LoginService {
-    constructor(private principal: Principal, private trackerService: JhiTrackerService, private authServerProvider: AuthServerProvider) {}
+    constructor(
+        private accountService: AccountService,
+        private trackerService: JhiTrackerService,
+        private authServerProvider: AuthServerProvider
+    ) {}
 
     login(credentials, callback?) {
         const cb = callback || function() {};
@@ -14,7 +18,7 @@ export class LoginService {
         return new Promise((resolve, reject) => {
             this.authServerProvider.login(credentials).subscribe(
                 data => {
-                    this.principal.identity(true).then(account => {
+                    this.accountService.identity(true).then(account => {
                         this.trackerService.sendActivity();
                         resolve(data);
                     });
@@ -34,7 +38,6 @@ export class LoginService {
     }
 
     logout() {
-        this.authServerProvider.logout().subscribe();
-        this.principal.authenticate(null);
+        this.authServerProvider.logout().subscribe(null, null, () => this.accountService.authenticate(null));
     }
 }
