@@ -1,5 +1,5 @@
 import { Directive, Input, TemplateRef, ViewContainerRef } from '@angular/core';
-import { Principal } from 'app/core/auth/principal.service';
+import { AccountService } from 'app/core/auth/account.service';
 
 /**
  * @whatItDoes Conditionally includes an HTML element if current user has any
@@ -18,22 +18,25 @@ import { Principal } from 'app/core/auth/principal.service';
 export class HasAnyAuthorityDirective {
     private authorities: string[];
 
-    constructor(private principal: Principal, private templateRef: TemplateRef<any>, private viewContainerRef: ViewContainerRef) {}
+    constructor(
+        private accountService: AccountService,
+        private templateRef: TemplateRef<any>,
+        private viewContainerRef: ViewContainerRef
+    ) {}
 
     @Input()
     set jhiHasAnyAuthority(value: string | string[]) {
         this.authorities = typeof value === 'string' ? [value] : value;
         this.updateView();
         // Get notified each time authentication state changes.
-        this.principal.getAuthenticationState().subscribe(identity => this.updateView());
+        this.accountService.getAuthenticationState().subscribe(identity => this.updateView());
     }
 
     private updateView(): void {
-        this.principal.hasAnyAuthority(this.authorities).then(result => {
-            this.viewContainerRef.clear();
-            if (result) {
-                this.viewContainerRef.createEmbeddedView(this.templateRef);
-            }
-        });
+        const hasAnyAuthority = this.accountService.hasAnyAuthority(this.authorities);
+        this.viewContainerRef.clear();
+        if (hasAnyAuthority) {
+            this.viewContainerRef.createEmbeddedView(this.templateRef);
+        }
     }
 }
