@@ -1,2 +1,2 @@
 export const DATE_FORMAT = 'YYYY-MM-DD';
-export const DATE_TIME_FORMAT = 'YYYY-MM-DDThh:mm';
+export const DATE_TIME_FORMAT = 'YYYY-MM-DDTHH:mm';
