@@ -9,7 +9,7 @@ import { Moment } from 'moment';
 @Injectable()
 export class NgbDateMomentAdapter extends NgbDateAdapter<Moment> {
     fromModel(date: Moment): NgbDateStruct {
-        if (date != null && date.isValid()) {
+        if (date != null && moment.isMoment(date) && date.isValid()) {
             return { year: date.year(), month: date.month() + 1, day: date.date() };
         }
         return null;
