@@ -6,53 +6,42 @@ import { SERVER_API_URL } from 'app/app.constants';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 
 describe('Service Tests', () => {
-    describe('Logs Service', () => {
-        let service: LogsService;
-        let httpMock;
+  describe('Logs Service', () => {
+    let service: LogsService;
+    let httpMock;
+    let expectedResult;
 
-        beforeEach(() => {
-            TestBed.configureTestingModule({
-                imports: [HttpClientTestingModule]
-            });
+    beforeEach(() => {
+      TestBed.configureTestingModule({
+        imports: [HttpClientTestingModule]
+      });
 
-            service = TestBed.get(LogsService);
-            httpMock = TestBed.get(HttpTestingController);
-        });
-
-        afterEach(() => {
-            httpMock.verify();
-        });
-
-        describe('Service methods', () => {
-            it('should call correct URL', () => {
-                service.findAll().subscribe(() => {});
-
-                const req = httpMock.expectOne({ method: 'GET' });
-                const resourceUrl = SERVER_API_URL + 'management/logs';
-                expect(req.request.url).toEqual(resourceUrl);
-            });
-
-            it('should return Logs', () => {
-                const log = new Log('main', 'ERROR');
-
-                service.findAll().subscribe(received => {
-                    expect(received.body[0]).toEqual(log);
-                });
-
-                const req = httpMock.expectOne({ method: 'GET' });
-                req.flush([log]);
-            });
-
-            it('should change log level', () => {
-                const log = new Log('main', 'ERROR');
-
-                service.changeLevel(log).subscribe(received => {
-                    expect(received.body[0]).toEqual(log);
-                });
-
-                const req = httpMock.expectOne({ method: 'PUT' });
-                req.flush([log]);
-            });
-        });
+      expectedResult = {};
+      service = TestBed.get(LogsService);
+      httpMock = TestBed.get(HttpTestingController);
     });
+
+    afterEach(() => {
+      httpMock.verify();
+    });
+
+    describe('Service methods', () => {
+      it('should call correct URL', () => {
+        service.findAll().subscribe();
+
+        const req = httpMock.expectOne({ method: 'GET' });
+        const resourceUrl = SERVER_API_URL + 'management/loggers';
+        expect(req.request.url).toEqual(resourceUrl);
+      });
+
+      it('should change log level', () => {
+        service.changeLevel('main', 'ERROR').subscribe();
+
+        const req = httpMock.expectOne({ method: 'POST' });
+        const resourceUrl = SERVER_API_URL + 'management/loggers/main';
+        expect(req.request.url).toEqual(resourceUrl);
+        expect(req.request.body).toEqual({ configuredLevel: 'ERROR' });
+      });
+    });
+  });
 });
