@@ -3,10 +3,12 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import Spy = jasmine.Spy;
 
 export class MockActiveModal extends SpyObject {
-    dismissSpy: Spy;
+  dismissSpy: Spy;
+  closeSpy: Spy;
 
-    constructor() {
-        super(NgbActiveModal);
-        this.dismissSpy = this.spy('dismiss').andReturn(this);
-    }
+  constructor() {
+    super(NgbActiveModal);
+    this.dismissSpy = this.spy('dismiss').andReturn(this);
+    this.closeSpy = this.spy('close').andReturn(this);
+  }
 }
