@@ -3,10 +3,10 @@ import { JhiEventManager } from 'ng-jhipster';
 import Spy = jasmine.Spy;
 
 export class MockEventManager extends SpyObject {
-    broadcastSpy: Spy;
+  broadcastSpy: Spy;
 
-    constructor() {
-        super(JhiEventManager);
-        this.broadcastSpy = this.spy('broadcast').andReturn(this);
-    }
+  constructor() {
+    super(JhiEventManager);
+    this.broadcastSpy = this.spy('broadcast').andReturn(this);
+  }
 }
