@@ -2,9 +2,9 @@ import { SpyObject } from './spyobject';
 import { JhiTrackerService } from 'app/core/tracker/tracker.service';
 
 export class MockTrackerService extends SpyObject {
-    constructor() {
-        super(JhiTrackerService);
-    }
+  constructor() {
+    super(JhiTrackerService);
+  }
 
-    connect() {}
+  connect() {}
 }
