@@ -5,68 +5,63 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { NgbActiveModal, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { JhiDataUtils, JhiDateUtils, JhiEventManager, JhiAlertService, JhiParseLinks } from 'ng-jhipster';
 
-import { Principal, AccountService, LoginModalService, JhiTrackerService } from 'app/core';
-import { MockPrincipal } from './helpers/mock-principal.service';
+import { AccountService, LoginModalService, JhiTrackerService } from 'app/core';
 import { MockAccountService } from './helpers/mock-account.service';
 import { MockActivatedRoute, MockRouter } from './helpers/mock-route.service';
 import { MockActiveModal } from './helpers/mock-active-modal.service';
 import { MockEventManager } from './helpers/mock-event-manager.service';
 
 @NgModule({
-    providers: [
-        DatePipe,
-        JhiDataUtils,
-        JhiDateUtils,
-        JhiParseLinks,
-        {
-            provide: JhiTrackerService,
-            useValue: null
-        },
-        {
-            provide: JhiEventManager,
-            useClass: MockEventManager
-        },
-        {
-            provide: NgbActiveModal,
-            useClass: MockActiveModal
-        },
-        {
-            provide: ActivatedRoute,
-            useValue: new MockActivatedRoute({ id: 123 })
-        },
-        {
-            provide: Router,
-            useClass: MockRouter
-        },
-        {
-            provide: Principal,
-            useClass: MockPrincipal
-        },
-        {
-            provide: AccountService,
-            useClass: MockAccountService
-        },
-        {
-            provide: LoginModalService,
-            useValue: null
-        },
-        {
-            provide: ElementRef,
-            useValue: null
-        },
-        {
-            provide: Renderer,
-            useValue: null
-        },
-        {
-            provide: JhiAlertService,
-            useValue: null
-        },
-        {
-            provide: NgbModal,
-            useValue: null
-        }
-    ],
-    imports: [HttpClientTestingModule]
+  providers: [
+    DatePipe,
+    JhiDataUtils,
+    JhiDateUtils,
+    JhiParseLinks,
+    {
+      provide: JhiTrackerService,
+      useValue: null
+    },
+    {
+      provide: JhiEventManager,
+      useClass: MockEventManager
+    },
+    {
+      provide: NgbActiveModal,
+      useClass: MockActiveModal
+    },
+    {
+      provide: ActivatedRoute,
+      useValue: new MockActivatedRoute({ id: 123 })
+    },
+    {
+      provide: Router,
+      useClass: MockRouter
+    },
+    {
+      provide: AccountService,
+      useClass: MockAccountService
+    },
+    {
+      provide: LoginModalService,
+      useValue: null
+    },
+    {
+      provide: ElementRef,
+      useValue: null
+    },
+    {
+      provide: Renderer,
+      useValue: null
+    },
+    {
+      provide: JhiAlertService,
+      useValue: null
+    },
+    {
+      provide: NgbModal,
+      useValue: null
+    }
+  ],
+  imports: [HttpClientTestingModule]
 })
 export class ZeusTestModule {}
