@@ -6,6 +6,7 @@ const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
 const SimpleProgressWebpackPlugin = require('simple-progress-webpack-plugin');
 const path = require('path');
+const sass = require('sass');
 
 const utils = require('./utils.js');
 const commonConfig = require('./webpack.common.js');
@@ -54,13 +55,13 @@ module.exports = (options) => webpackMerge(commonConfig({ env: ENV }), {
         rules: [{
             test: /\.ts$/,
             enforce: 'pre',
-            loaders: 'tslint-loader',
-            exclude: ['node_modules', new RegExp('reflect-metadata\\' + path.sep + 'Reflect\\.ts')]
+            loader: 'tslint-loader',
+            exclude: [/(node_modules)/, new RegExp('reflect-metadata\\' + path.sep + 'Reflect\\.ts')]
         },
             {
                 test: /\.ts$/,
                 use: [
-                    { loader: 'angular2-template-loader' },
+                    'angular2-template-loader',
                     {
                         loader: 'cache-loader',
                         options: {
@@ -81,27 +82,24 @@ module.exports = (options) => webpackMerge(commonConfig({ env: ENV }), {
                             happyPackMode: true
                         }
                     },
-                    { loader: 'angular-router-loader' }
+                    'angular-router-loader'
                 ],
-                exclude: ['node_modules']
+                exclude: /(node_modules)/
             },
             {
                 test: /\.scss$/,
-                loaders: ['to-string-loader', 'css-loader', 'sass-loader'],
+                use: ['to-string-loader', 'css-loader', {
+                    loader: 'sass-loader',
+                    options: { implementation: sass }
+                }],
                 exclude: /(vendor\.scss|global\.scss)/
             },
             {
                 test: /(vendor\.scss|global\.scss)/,
-                loaders: ['style-loader', 'css-loader', 'postcss-loader', 'sass-loader']
-            },
-            {
-                test: /\.css$/,
-                loaders: ['to-string-loader', 'css-loader'],
-                exclude: /(vendor\.css|global\.css)/
-            },
-            {
-                test: /(vendor\.css|global\.css)/,
-                loaders: ['style-loader', 'css-loader']
+                use: ['style-loader', 'css-loader', 'postcss-loader', {
+                    loader: 'sass-loader',
+                    options: { implementation: sass }
+                }]
             }]
     },
     stats: options.stats,
@@ -127,8 +125,8 @@ module.exports = (options) => webpackMerge(commonConfig({ env: ENV }), {
         ),
         new writeFilePlugin(),
         new webpack.WatchIgnorePlugin([
-            utils.root('test'),
-        ]),
+            utils.root('test')
+        ])
     ],
     mode: 'development'
 });
